+++ weixin-java-common/src/main/java/me/chanjar/weixin/common/util/http/RequestExecutor.java
@@ -1,8 +1,6 @@
 package me.chanjar.weixin.common.util.http;
 
 import me.chanjar.weixin.common.exception.WxErrorException;
-import org.apache.http.HttpHost;
-import org.apache.http.impl.client.CloseableHttpClient;
 
 import java.io.IOException;
 
@@ -12,16 +10,16 @@ import java.io.IOException;
  * @param <T> 返回值类型
  * @param <E> 请求参数类型
  */
-public interface RequestExecutor<T, E> {
+public interface RequestExecutor<T, H, P, E> {
 
   /**
-   * @param httpclient 传入的httpClient
+   * @param httpClient
    * @param httpProxy  http代理对象，如果没有配置代理则为空
    * @param uri        uri
    * @param data       数据
    * @throws WxErrorException
    * @throws IOException
    */
-  T execute(CloseableHttpClient httpclient, HttpHost httpProxy, String uri, E data) throws WxErrorException, IOException;
+  T execute(H httpClient, P httpProxy, String uri, E data) throws WxErrorException, IOException;
 
 }
