+++ weixin-java-common/src/main/java/me/chanjar/weixin/common/util/http/apache/InputStreamResponseHandler.java
@@ -1,6 +1,5 @@
-package me.chanjar.weixin.common.util.http;
+package me.chanjar.weixin.common.util.http.apache;
 
-import org.apache.http.Consts;
 import org.apache.http.HttpEntity;
 import org.apache.http.HttpResponse;
 import org.apache.http.StatusLine;
@@ -9,25 +8,21 @@ import org.apache.http.client.ResponseHandler;
 import org.apache.http.util.EntityUtils;
 
 import java.io.IOException;
+import java.io.InputStream;
 
-/**
- * copy from {@link org.apache.http.impl.client.BasicResponseHandler}
- *
- * @author Daniel Qian
- */
-public class Utf8ResponseHandler implements ResponseHandler<String> {
+public class InputStreamResponseHandler implements ResponseHandler<InputStream> {
 
-  public static final ResponseHandler<String> INSTANCE = new Utf8ResponseHandler();
+  public static final ResponseHandler<InputStream> INSTANCE = new InputStreamResponseHandler();
 
   @Override
-  public String handleResponse(final HttpResponse response) throws IOException {
+  public InputStream handleResponse(final HttpResponse response) throws IOException {
     final StatusLine statusLine = response.getStatusLine();
     final HttpEntity entity = response.getEntity();
     if (statusLine.getStatusCode() >= 300) {
       EntityUtils.consume(entity);
       throw new HttpResponseException(statusLine.getStatusCode(), statusLine.getReasonPhrase());
     }
-    return entity == null ? null : EntityUtils.toString(entity, Consts.UTF_8);
+    return entity == null ? null : entity.getContent();
   }
 
 }
