+++ weixin-java-common/src/main/java/me/chanjar/weixin/common/util/http/jodd/MediaUploadRequestExecutor.java
@@ -1,12 +1,13 @@
 package me.chanjar.weixin.common.util.http.jodd;
 
+import jodd.http.HttpConnectionProvider;
 import jodd.http.HttpRequest;
 import jodd.http.HttpResponse;
 import jodd.http.ProxyInfo;
-import jodd.http.net.SocketHttpConnectionProvider;
 import me.chanjar.weixin.common.bean.result.WxError;
 import me.chanjar.weixin.common.bean.result.WxMediaUploadResult;
 import me.chanjar.weixin.common.exception.WxErrorException;
+import me.chanjar.weixin.common.util.http.RequestExecutor;
 
 import java.io.File;
 import java.io.IOException;
@@ -16,16 +17,15 @@ import java.io.IOException;
  *
  * @author Daniel Qian
  */
-public class MediaUploadRequestExecutor implements RequestExecutor<WxMediaUploadResult, File> {
+public class MediaUploadRequestExecutor implements RequestExecutor<WxMediaUploadResult, HttpConnectionProvider, ProxyInfo, File> {
 
   @Override
-  public WxMediaUploadResult execute(ProxyInfo httpProxy, String uri, File file) throws WxErrorException, IOException {
+  public WxMediaUploadResult execute(HttpConnectionProvider provider, ProxyInfo httpProxy, String uri, File file) throws WxErrorException, IOException {
     HttpRequest request = HttpRequest.post(uri);
     if (httpProxy != null) {
-      SocketHttpConnectionProvider provider = new SocketHttpConnectionProvider();
       provider.useProxy(httpProxy);
-      request.withConnectionProvider(provider);
     }
+    request.withConnectionProvider(provider);
     request.form("media", file);
     HttpResponse response = request.send();
     String responseContent =response.bodyText();
