+++ weixin-java-common/src/main/java/me/chanjar/weixin/common/util/http/jodd/SimplePostRequestExecutor.java
@@ -1,11 +1,12 @@
 package me.chanjar.weixin.common.util.http.jodd;
 
+import jodd.http.HttpConnectionProvider;
 import jodd.http.HttpRequest;
 import jodd.http.HttpResponse;
 import jodd.http.ProxyInfo;
-import jodd.http.net.SocketHttpConnectionProvider;
 import me.chanjar.weixin.common.bean.result.WxError;
 import me.chanjar.weixin.common.exception.WxErrorException;
+import me.chanjar.weixin.common.util.http.RequestExecutor;
 
 import java.io.IOException;
 
@@ -14,16 +15,15 @@ import java.io.IOException;
  *
  * @author Daniel Qian
  */
-public class SimplePostRequestExecutor implements RequestExecutor<String, String> {
+public class SimplePostRequestExecutor implements RequestExecutor<String, HttpConnectionProvider, ProxyInfo, String> {
 
   @Override
-  public String execute(ProxyInfo httpProxy, String uri, String postEntity) throws WxErrorException, IOException {
+  public String execute(HttpConnectionProvider provider, ProxyInfo httpProxy, String uri, String postEntity) throws WxErrorException, IOException {
     HttpRequest request = HttpRequest.post(uri);
     if (httpProxy != null) {
-      SocketHttpConnectionProvider provider = new SocketHttpConnectionProvider();
       provider.useProxy(httpProxy);
-      request.withConnectionProvider(provider);
     }
+    request.withConnectionProvider(provider);
     if (postEntity != null) {
       request.bodyText(postEntity);
     }
